refactor(playback): move multi-frame playback into play-view drawer; delete /play route (#2232)
The standalone /play/[climb_uuid] route duplicated nearly everything the existing
play-view drawer already does (board renderer, swipe nav, action bar, comments,
queue navigation). The only thing it owned uniquely was multi-frame playback:
the engine, BLE frame loop, peer sync, and on-screen controls.

This collapses the two surfaces. The drawer now wires usePlaybackEngine,
useClimbFrames, the latest-wins BLE drain (gated on isOpen so frames don't
bleed to the wall on close), and the PlaybackStateChanged subscribe/publish
loop. SwipeBoardCarousel renders the engine's current frame on routes and the
static frames string on single-frame climbs; zoomResetKey pins to climb UUID
so pinch zoom survives frame ticks. PlaybackControls slot below the action bar
when isAnimatable.

Old /play/{uuid} URLs (canonical and /b/ short-slug shapes) 301-redirect to
the equivalent /view/{slug-uuid} so bookmarks keep working.

Strips the corresponding glue: queue-control-bar isPlayPage branches and
desktop "Open in Full" link, queue-nav-button navigate prop, board-page header
back button and isPlayPage angle gate, use-resolved-board-details isPlayPage
return field, constructPlayUrlWithSlugs / tryConstructSlugPlayUrl /
constructBoardSlugPlayUrl in url-utils, metadata.play.* + playView.noClimbSelected
+ playView.browseClimbs + queueBar.ariaLabels.enterPlayMode + header.backToList
i18n keys across en-US/es/fr.

Existing tests that mocked /play/ pathnames switch to /view/ to keep covering
the now-canonical surface. The playback engine + backend mutation contract are
already covered by use-playback-engine.test.ts and publish-playback-state.test.ts.
