--- packages/web/app/[board_name]/[layout_id]/[size_id]/[set_ids]/[angle]/play/[climb_uuid]/play-view.module.css
@@ -1,71 +1,0 @@
-/* Layout-only styles - colors are applied via inline styles using themeTokens */
-
-.pageContainer {
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  overflow-y: auto;
-  touch-action: manipulation;
-}
-
-.contentWrapper {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  overflow: hidden;
-  position: relative;
-  min-height: 0;
-}
-
-.climbTitleContainer {
-  width: 100%;
-  flex-shrink: 0;
-}
-
-.swipeContainer {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: flex-start;
-  touch-action: manipulation;
-  overflow: hidden;
-  min-height: 0;
-  position: relative;
-}
-
-.boardContainer {
-  flex: 1;
-  display: flex;
-  align-items: flex-start;
-  justify-content: center;
-  padding: 0;
-  min-height: 0;
-  max-height: 100%;
-  width: 100%;
-}
-
-.emptyState {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  gap: 16px;
-  padding: 24px;
-  text-align: center;
-}
-
-.extrasSection {
-  flex-shrink: 0;
-  border-top: 1px solid var(--neutral-200);
-  padding-top: 8px;
-}
-
-/* Desktop adjustments */
-@media (min-width: 768px) {
-  .boardContainer {
-    max-width: 600px;
-    margin: 0 auto;
-  }
-}
