--- packages/web/app/[board_name]/[layout_id]/[size_id]/[set_ids]/[angle]/play/layout-client.module.css
@@ -1,66 +1,0 @@
-.playLayout {
-  background: transparent;
-  height: 100%;
-  position: relative;
-  overflow-x: hidden;
-  max-width: 100%;
-}
-
-.mainContent {
-  background: transparent;
-  padding: 0;
-  margin-right: 0;
-  transition: margin-right 0.2s;
-  overflow-x: hidden;
-  max-width: 100%;
-  display: flex;
-  flex-direction: column;
-}
-
-.sider {
-  position: fixed !important;
-  right: 0;
-  top: var(--global-header-height);
-  height: calc(100dvh - var(--global-header-height) - 64px);
-  overflow-y: auto;
-  background: var(--semantic-surface);
-  box-shadow: -2px 0 8px rgba(0, 0, 0, 0.08);
-  display: none;
-  padding: 0 8px 20px 8px;
-}
-
-.sidebarContent {
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-}
-
-.sidebarHeader {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 12px 8px 8px 8px;
-  border-bottom: 1px solid var(--neutral-200);
-}
-
-.sidebarTitle {
-  margin: 0;
-  font-size: 16px;
-  font-weight: 500;
-}
-
-.queueListWrapper {
-  flex: 1;
-  overflow: auto;
-}
-
-/* Only show sider on desktop */
-@media (min-width: 768px) {
-  .sider {
-    display: block;
-  }
-
-  .mainContent {
-    margin-right: 400px; /* Same as sider width */
-  }
-}
