--- packages/web/app/[board_name]/[layout_id]/[size_id]/[set_ids]/[angle]/play/layout.tsx
@@ -1,20 +1,0 @@
-import React, { type PropsWithChildren } from 'react';
-
-import type { BoardRouteParameters } from '@/app/lib/types';
-import { parseRouteParams } from '@/app/lib/url-utils.server';
-import { getBoardDetailsForBoard } from '@/app/lib/board-utils';
-import PlayLayoutClient from './layout-client';
-
-type LayoutProps = {
-  params: Promise<BoardRouteParameters>;
-};
-
-export default async function PlayLayout(props: PropsWithChildren<LayoutProps>) {
-  const params = await props.params;
-  const { children } = props;
-
-  const { parsedParams } = await parseRouteParams(params);
-  const boardDetails = getBoardDetailsForBoard(parsedParams);
-
-  return <PlayLayoutClient boardDetails={boardDetails}>{children}</PlayLayoutClient>;
-}
