+++ packages/web/app/components/queue-control/queue-nav-button.tsx
@@ -3,9 +3,7 @@
 import React, { useCallback } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useQueueActions, useSessionData } from '../graphql-queue';
-import { constructPlayUrlWithSlugs } from '@/app/lib/url-utils';
 import type { BoardDetails } from '@/app/lib/types';
-import { useResolvedBoardDetails } from '@/app/hooks/use-resolved-board-details';
 import FastForwardOutlined from '@mui/icons-material/FastForwardOutlined';
 import FastRewindOutlined from '@mui/icons-material/FastRewindOutlined';
 import { track } from '@/app/lib/analytics';
@@ -15,7 +13,6 @@ type Direction = 'next' | 'previous';
 
 type QueueNavButtonProps = {
   direction: Direction;
-  navigate: boolean;
   boardDetails?: BoardDetails;
 };
 
@@ -30,37 +27,18 @@ const ICON = { next: FastForwardOutlined, previous: FastRewindOutlined } as cons
  * and no driver transfer; the presser stays a non-driver, only the shared
  * queue position moves. `setCurrentClimbQueueItem` updates the queue state
  * (and broadcasts via the persistent-session subscription when party is
- * active) but never calls `takeControl`.
+ * active) but never calls `takeControl`. On `/view/` pages the drawer's
+ * `useDrawerUrlSync` owns URL state, so this button never touches history.
  */
-export default function QueueNavButton({ direction, navigate, boardDetails }: QueueNavButtonProps) {
+export default function QueueNavButton({ direction, boardDetails }: QueueNavButtonProps) {
   const { t } = useTranslation('climbs');
   // Statically-resolvable t() calls so the i18n linter can verify both keys.
   const ariaLabel = direction === 'next' ? t('actions.navigation.nextClimb') : t('actions.navigation.previousClimb');
   const Icon = ICON[direction];
   const { setCurrentClimbQueueItem, getNextClimbQueueItem, getPreviousClimbQueueItem } = useQueueActions();
   const { viewOnlyMode, isPersistentSessionActive, isDriver } = useSessionData();
-  const { rawParams, angle, searchParams, isPlayPage, resolvedDetails } = useResolvedBoardDetails(boardDetails);
 
   const target = direction === 'next' ? getNextClimbQueueItem() : getPreviousClimbQueueItem();
-
-  const buildPlayUrl = () => {
-    if (!target) return '';
-    const slugUrl =
-      resolvedDetails.layout_name && resolvedDetails.size_name && resolvedDetails.set_names
-        ? constructPlayUrlWithSlugs(
-            resolvedDetails.board_name,
-            resolvedDetails.layout_name,
-            resolvedDetails.size_name,
-            resolvedDetails.size_description,
-            resolvedDetails.set_names,
-            angle,
-            target.climb.uuid,
-            target.climb.name,
-          )
-        : `/${rawParams.board_name}/${rawParams.layout_id}/${rawParams.size_id}/${rawParams.set_ids}/${rawParams.angle}/play/${target.climb.uuid}`;
-    const queryString = searchParams.toString();
-    return queryString ? `${slugUrl}?${queryString}` : slugUrl;
-  };
 
   const fireAdvance = useCallback(() => {
     if (!target || viewOnlyMode) return;
@@ -89,20 +67,8 @@ export default function QueueNavButton({ direction, navigate, boardDetails }: Qu
       mode: isPersistentSessionActive ? 'party' : 'solo',
       boardLayout,
     });
-    // On /play/ routes the URL is the source of truth — push the new climb's
-    // play URL. On /view/ routes useDrawerUrlSync handles the URL via the
-    // drawer's open lifecycle; pushing here would race the hook's cleanup.
-    if (navigate && isPlayPage) {
-      const url = buildPlayUrl();
-      if (url) window.history.pushState(null, '', url);
-    }
-    // buildPlayUrl is a closure over current state; recreating each fire is
-    // intentional. target is the only ref-stable input.
-    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [
     target,
-    navigate,
-    isPlayPage,
     viewOnlyMode,
     setCurrentClimbQueueItem,
     boardDetails?.layout_name,
